back/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
cryptography==38.0.3
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-passwordreset==1.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.10
ndg-httpsclient==0.5.1
Pillow==9.3.0
psycopg2==2.9.5
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stripe==2.55.2
urllib3==1.26.3
whitenoise==5.2.0
